pyproject.toml
[project]
name = "TrainyBackEnd"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "AnthonyLoche", email = "anthonylocheifc@gmail.com"},
]
dependencies = ["django>=5.1.6", "djangorestframework>=3.15.2", "pydotplus>=2.0.2", "django-extensions>=3.2.3", "faker>=36.1.1", "djangorestframework-simplejwt>=5.5.0", "drf-spectacular>=0.28.0", "cloudinary>=1.42.2", "python-dotenv>=1.0.1", "Pillow>=11.1.0", "python-magic>=0.4.27"]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.pdm]
distribution = true

[tool.pdm.scripts]
dev = "python manage.py runserver"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
changepassword = "python manage.py changepassword"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
migrate_db = "python manage.py migrate_db"
shell = "python manage.py shell"
test = "python manage.py test"
startapp = "python manage.py startapp {args}"
loaddata = "python manage.py loaddata {args}"
dumpdata = "python manage.py dumpdata {args}"
model = "python manage.py graph_models -a -g -o models.png"
req = "pdm export -f requirements --without-hashes -o requirements.txt"
